Executable enabled by OrionI.service
This is our executable file for the OrionI.service in systemd/system ie our executable bash file to run on boot. Currently, this file calls a python script in the Pi's home directory 'test1.py' which turns on two GPIO pins, and turns them off aftrer 10 seconds. This was used as a test, LEDs were hooked up to ensure that the service runs on boot (it does!)

The end goal for this executable file is to contain instructions which need to be run every time the Jeep turns on, so we can enable the CAN system, update our Pi, and pull up the Python UI without the user having to do anything. This gives the user a sense of a purpose built OS/platform, in which they don't have to launch anything themselves.